Change Rect getters to use (r *Rect)
I had the wrong assumption that whenever coding a function to a given
struct, if that function would change the struct, then it will require a
pointer to that associated struct:

    func (r *Rect) SetWidth(w float64)

While, also following this assumption, a function that doesn't change
the underlying structure wouldn't require a pointer:
    func (r Rect) GetWidth() float64

HOWEVER, above assumpton was wrong. The GetWidth declaration above leads
to an issue while passing that funcion as a parameter.

In the above assumption, the passed function would always point to a
snapshot of the Rect struct.

This means, whenever the Rect struct changes, a previous reference to
GetWidth will always return the width value present at the time GetWidth
funciton was retreived:
    r := Rect{w: 3.0, h:4.0}
    doSomething(r.SetWidth, r.getWidth);

    func doSomething(setter func(float64), getter func() float64) {
        getter() // this returns 3.0

        setter(1.5) // this makes r.width = 1.5

        getter() // this returns 3.0, not 1.5
    }

By declarng getWidth as follows:
    func (r *Rect) GetWidth() float64

Then:

    func doSomething(setter func(float64), getter func() float64) {
        getter() // this returns 3.0

        setter(1.5) // this makes r.width = 1.5

        getter() // this returns 1.5 as expected
    }